+++ src/Fields/Repeater.php
@@ -2,41 +2,15 @@
 
 namespace Towa\Acf\Fields;
 
-use Towa\Acf\BaseField;
+use Towa\Acf\BaseParentField;
 
-class Repeater extends BaseField
+class Repeater extends BaseParentField
 {
     protected $prefix;
     protected $name = 'repeater';
     protected $type = 'repeater';
     protected $label = 'Wiederholung';
-
-    /**
-     * @param string $name_key
-     * @param string $label
-     * @param array  $parameter
-     *
-     * @return array
-     */
-    public function build(array $parameter = [])
-    {
-
-//		$this->set_name( $name_key );
-
-        $add_to_defaults = [
-            'button_label' => 'hinzufügen',
-            'sub_fields'   => [],
-        ];
-
-        //		$add_to_defaults = array_merge( (array) $add_to_defaults, (array) $parameter );
-
-        //		foreach ( $add_to_defaults['sub_fields'] as &$value ) {
-
-        //TODO change key ?
-        //			$value['key'] = $value['key'] . '_' . $this->get_name();
-        //		}
-
-        //		return parent::build( $add_to_defaults );
-        return parent::build(array_merge((array) $add_to_defaults, (array) $parameter));
-    }
+    protected $args = [
+        'button_label' => 'hinzufügen',
+    ];
 }